requirements.txt
# Core Dependencies
fastapi==0.109.0
uvicorn==0.27.0
streamlit==1.31.0
python-multipart==0.0.6

# AI/ML Services
mistralai==0.1.3
elevenlabs==0.2.27
openai-whisper>=20240101
transformers==4.37.0
torch==2.1.2

# Audio Processing
pydub==0.25.1
soundfile==0.12.1
librosa==0.10.1

# Data Processing & Fine-tuning
datasets==4.6.1
dill==0.4.0
multiprocess==0.70.18
xxhash==3.6.0
pyarrow==23.0.1

# Monitoring & Tracking
wandb>=0.16.4
weave==0.50.0

# Database
sqlalchemy==2.0.25

# Blockchain
web3==6.11.3

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
requests==2.32.5
numpy==1.26.3
PyJWT==2.10.1
email-validator==2.2.0

